+++ src/commands/validate.ts
@@ -2,8 +2,8 @@ import { Arguments, CommandBuilder } from "yargs";
 
 import { authenticate, BaseOptions, baseOptions, printError } from "../shared";
 import spinner from "../services/spinner/index";
-import handlers from "./push/handlers/handlers";
-import { UserVariableInputs } from "./push/handlers/validators";
+import handlers from "./deploy/handlers/handlers";
+import { UserVariableInputs } from "./deploy/handlers/validators";
 
 export interface Options extends BaseOptions {
   config?: string;