+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+- Renamed push command to deploy
+
 ### Removed
 - Deleted alerts commands
 - Deleted destroy command